### 2018-07-09 #817 @whedon
[PRE REVIEW]: GEM: A Python package for graph embedding methods
**Submitting author:** @palash1992 (<a href="http://orcid.org/0000-0003-2455-2160">Palash Goyal</a>)
**Repository:** <a href="https://github.com/palash1992/GEM" target ="_blank">https://github.com/palash1992/GEM</a>
**Version:** v1.0.0
