### 2018-07-10 #817 @danielskatz
@palash1992
1. Can you clean up the references in your paper?  A bunch need DOIs, and have case problems, eg ACM as Acm.
1. Can you suggest any reviewers?  [This list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission, or you may have other suggestions for people from the field of your work.
