### 2018-07-10 cce1fee
Modified citations
