+++ paper.bib
@@ -1,7 +1,8 @@
 @article{roweis2000nonlinear,
   title={Nonlinear dimensionality reduction by locally linear embedding},
   author={Roweis, Sam T and Saul, Lawrence K},
-  journal={science},
+  journal={Science},
+  doi={10.1126/science.290.5500.2323},
   volume={290},
   number={5500},
   pages={2323--2326},
@@ -12,7 +13,7 @@
 @inproceedings{belkin2002laplacian,
   title={Laplacian eigenmaps and spectral techniques for embedding and clustering},
   author={Belkin, Mikhail and Niyogi, Partha},
-  booktitle={Advances in neural information processing systems},
+  booktitle={Advances in Neural Information Processing System},
   pages={585--591},
   year={2002}
 }
@@ -20,7 +21,8 @@
 @inproceedings{ahmed2013distributed,
   title={Distributed large-scale natural graph factorization},
   author={Ahmed, Amr and Shervashidze, Nino and Narayanamurthy, Shravan and Josifovski, Vanja and Smola, Alexander J},
-  booktitle={Proceedings of the 22nd international conference on World Wide Web},
+  booktitle={Proceedings of the 22nd International Conference on World Wide Web},
+  doi={10.1145/2488388.2488393},
   pages={37--48},
   year={2013},
   organization={ACM}
@@ -29,7 +31,8 @@
 @inproceedings{ou2016asymmetric,
   title={Asymmetric transitivity preserving graph embedding},
   author={Ou, Mingdong and Cui, Peng and Pei, Jian and Zhang, Ziwei and Zhu, Wenwu},
-  booktitle={Proceedings of the 22nd ACM SIGKDD international conference on Knowledge discovery and data mining},
+  booktitle={Proceedings of the 22nd ACM SIGKDD International Conference on Knowledge Discovery and Data Mining},
+  doi={10.1145/2939672.2939751},
   pages={1105--1114},
   year={2016},
   organization={ACM}
@@ -38,7 +41,8 @@
 @inproceedings{wang2016structural,
   title={Structural deep network embedding},
   author={Wang, Daixin and Cui, Peng and Zhu, Wenwu},
-  booktitle={Proceedings of the 22nd ACM SIGKDD international conference on Knowledge discovery and data mining},
+  booktitle={Proceedings of the 22nd ACM SIGKDD International Conference on Knowledge Discovery and Data Mining},
+  doi={10.1145/2939672.2939753},
   pages={1225--1234},
   year={2016},
   organization={ACM}
@@ -47,7 +51,8 @@
 @inproceedings{grover2016node2vec,
   title={node2vec: Scalable feature learning for networks},
   author={Grover, Aditya and Leskovec, Jure},
-  booktitle={Proceedings of the 22nd ACM SIGKDD international conference on Knowledge discovery and data mining},
+  booktitle={Proceedings of the 22nd ACM SIGKDD International Conference on Knowledge Discovery and Data Mining},
+  doi={10.1145/2939672.2939754},
   pages={855--864},
   year={2016},
   organization={ACM}
@@ -57,6 +62,7 @@
   title={Enhancing the network embedding quality with structural similarity},
   author={Lyu, Tianshu and Zhang, Yuan and Zhang, Yan},
   booktitle={Proceedings of the 2017 ACM on Conference on Information and Knowledge Management},
+  doi={10.1145/3132847.3132900},
   pages={147--156},
   year={2017},
   organization={ACM}
@@ -66,6 +72,7 @@
   title={Properties of Vector Embeddings in Social Networks},
   author={Salehi Rizi, Fatemeh and Granitzer, Michael and Ziegler, Konstantin},
   journal={Algorithms},
[…]
