### 2018-07-10 #817 @palash1992
@danielskatz

   1. I have cleaned up the references and modified the paper.bib file in
   Github. I added the missing DOIs (except 1 for which I could not find it).
   The problem about case you mentioned "Acm" is not in the bib but in the
   compiled PDF. I am not sure why that is the case. We can see if the problem
   persists and modify accordingly.
   2. Here are the reviewers who would be suitable for review:

   - jankatins
   - gabrielspadon
   - adebytes
   - paulgovan

Thanks! Let me know if there are any other concerns.

On Tue, Jul 10, 2018 at 1:41 AM, Daniel S. Katz <notifications@github.com>
wrote:

> @palash1992
> <https://urldefense.proofpoint.com/v2/url?u=https-3A__github.com_palash1992&d=DwMFaQ&c=clK7kQUTWtAVEOVIgvi0NU5BOUHhpN0H8p7CSfnc_gI&r=I_A-ZJGnHcHAuxuBmINQ-Q&m=UagMmUqlcX0Vv7jM-_Mpl8A2teKfEdMYXVkBXcz-EY0&s=V8xSto08YZtP9qSzjM94vEPr-UV5M5VcDdpHP-nirqY&e=>
>
>    1. Can you clean up the references in your paper? A bunch need DOIs,
>    and have case problems, eg ACM as Acm.
>    2. Can you suggest any reviewers? This list of people
>    <https://urldefense.proofpoint.com/v2/url?u=https-3A__bit.ly_joss-2Dreviewers&d=DwMFaQ&c=clK7kQUTWtAVEOVIgvi0NU5BOUHhpN0H8p7CSfnc_gI&r=I_A-ZJGnHcHAuxuBmINQ-Q&m=UagMmUqlcX0Vv7jM-_Mpl8A2teKfEdMYXVkBXcz-EY0&s=ZXVEmSP98HmAs45AH_kF6wf0MBhSibN5XxXAkOiaUW8&e=>
>    have already agreed to review for JOSS and may be suitable for this
>    submission, or you may have other suggestions for people from the field of
>    your work.
>
