### 2018-07-15 #817 @danielskatz
@palash1992 The Bib issues can be fixed by putting {}s around letters/words that should be kept in the case that they are.  Otherwise, the paper generation process may change them to what it thinks it correct.
