+++ paper.bib
@@ -31,7 +31,7 @@
 @inproceedings{ou2016asymmetric,
   title={Asymmetric transitivity preserving graph embedding},
   author={Ou, Mingdong and Cui, Peng and Pei, Jian and Zhang, Ziwei and Zhu, Wenwu},
-  booktitle={Proceedings of the 22nd ACM SIGKDD International Conference on Knowledge Discovery and Data Mining},
+  booktitle={Proceedings of the 22nd {ACM SIGKDD} International Conference on Knowledge Discovery and Data Mining},
   doi={10.1145/2939672.2939751},
   pages={1105--1114},
   year={2016},
@@ -41,7 +41,7 @@
 @inproceedings{wang2016structural,
   title={Structural deep network embedding},
   author={Wang, Daixin and Cui, Peng and Zhu, Wenwu},
-  booktitle={Proceedings of the 22nd ACM SIGKDD International Conference on Knowledge Discovery and Data Mining},
+  booktitle={Proceedings of the 22nd {ACM SIGKDD} International Conference on Knowledge Discovery and Data Mining},
   doi={10.1145/2939672.2939753},
   pages={1225--1234},
   year={2016},
@@ -51,7 +51,7 @@
 @inproceedings{grover2016node2vec,
   title={node2vec: Scalable feature learning for networks},
   author={Grover, Aditya and Leskovec, Jure},
-  booktitle={Proceedings of the 22nd ACM SIGKDD International Conference on Knowledge Discovery and Data Mining},
+  booktitle={Proceedings of the 22nd {ACM SIGKDD} International Conference on Knowledge Discovery and Data Mining},
   doi={10.1145/2939672.2939754},
   pages={855--864},
   year={2016},
@@ -61,7 +61,7 @@
 @inproceedings{lyu2017enhancing,
   title={Enhancing the network embedding quality with structural similarity},
   author={Lyu, Tianshu and Zhang, Yuan and Zhang, Yan},
-  booktitle={Proceedings of the 2017 ACM on Conference on Information and Knowledge Management},
+  booktitle={Proceedings of the 2017 {ACM} on Conference on Information and Knowledge Management},
   doi={10.1145/3132847.3132900},
   pages={147--156},
   year={2017},
