### 2018-07-16 #817 @palash1992
@danielskatz I have made the modification. Hopefully it should be fine now.
