### 2018-07-23 #817 @palash1992
@danielskatz : In that case, here are some other reviewers proficient in Python with focus machine learning: @rougier , @sealhuang, @ahurriyetoglu, @betatim and @jsgalan
