### 2018-07-23 #817 @danielskatz
👋  @rougier, @sealhuang, @ahurriyetoglu, @betatim, or @jsgalan - would any of you be willing to review this JOSS submission?
