### 2018-07-31 #817 @corybrunson
Yes, though i'll need a few weeks. I just agreed to review another paper closer to my areas of expertise/experience.
