### 2018-07-31 #817 @rougier
@danielskatz Sorry for the late answer, I was on vacation. If it's still time, I can also review.
