### 2018-08-04 #817 @corybrunson
@danielskatz i can review this coming week, but if @rougier is willing then i would defer to them (Python being their preferred language).
