### 2018-08-05 #817 @danielskatz
@rougier - thanks (and sorry for the delay in responding, I was teaching in a week-long summer school)
I'll add you and start the review with you and @jsgalan
@corybrunson  - thanks for volunteering as well, but 2 reviews is good for this one.
