### 2018-08-05 #876 @danielskatz
👋 @jsgalan & @rougier - thanks for agreeing to review this - please do so in this issue as described in the checklist and the second comment above.  Let me know if you have any questions or concerns.
