### 2018-08-05 149127c
Updated paper.md
