+++ paper.md
@@ -28,7 +28,7 @@ Many physical systems in the world involve interactions between different entiti
 
 ``GEM`` is a Python package which offers a general framework for graph embedding methods. It implements many state-of-the-art embedding techniques including Locally Linear Embedding [@roweis2000nonlinear], Laplacian Eigenmaps [@belkin2002laplacian], Graph Factorization [@ahmed2013distributed], HOPE [@ou2016asymmetric], SDNE [@wang2016structural] and node2vec [@grover2016node2vec]. It is formatted such that new methods can be easily added for comparison. Furthermore, the framework implements several functions to evaluate the quality of obtained embedding including graph reconstruction, link prediction, visualization and node classification. It supports many edge reconstruction metrics including cosine similarity, euclidean distance and decoder based. For node classification, it defaults to one-vs-rest logistic regression classifier and supports other classifiers. For faster execution, C++ backend is integrated using Boost for supported methods.
 
-``GEM`` was designed to be used by researchers studying graphs. It has already been used in a number of scientific publications to compare novel methods against the state-of-the-art and general evaluation [@salehi2017properties, @lyu2017enhancing]. A paper showcasing the results using ``GEM`` on various real world datasets can be accessed [goyal2017graph]. The source code of ``GEM`` is made available at <https://github.com/palash1992/GEM>. Bug reports and feedback can be directed to the Github issues page (<https://github.com/palash1992/GEM/issues>).
+``GEM`` was designed to be used by researchers studying graphs. It has already been used in a number of scientific publications to compare novel methods against the state-of-the-art and general evaluation [@salehi2017properties, @lyu2017enhancing]. A paper showcasing the results using ``GEM`` on various real world datasets can be accessed [@goyal2017graph]. The source code of ``GEM`` is made available at <https://github.com/palash1992/GEM>. Bug reports and feedback can be directed to the Github issues page (<https://github.com/palash1992/GEM/issues>).
 
 
 # References
