### 2018-08-05 #876 @palash1992
@danielskatz Thanks for pointing it out. I have modified the manuscript to reflect the change.
