### 2018-08-05 #876 @jsgalan
Hi all,
I just finished reviewing this interesting and novel package.

After installing and revising, a few comments mostly Unrelated on the implementation
- I tested it under OSX, Python 2.7.9 on the Enthought Canopy Environment
- All incompatibilities were cleared and all dependencies through PiP in the Python Canopy Terminal
- Thean, gave me a problem importing "_gof_", so following forum instructions had to uninstall and re-installed ([The Bleeding Edge version](http://deeplearning.net/software/theano/install.html#bleeding-edge-install-instructions)) through cloning the git repository

Even though is a Theano maybe should be added as a warning of the package

- From the example, i had an error on this block
> edge_f = 'gem/data/karate.edgelist'
> isDirected = True

I just got the file and place it in the working directory.  Should this dataset be placed inside the package itself under gem.datasets.karate?

Also, Ifound the following errors:
- when trying node2vec method

> _[Errno 2] No such file or directory
> Exception                                 Traceback (most recent call last)
> /JOSS/GEM/test1gem.py in <module>()
>      42     t1 = time()
>      43     # Learn embedding - accepts a networkx graph or file with edge list
> ---> 44     Y, t = embedding.learn_embedding(graph=G, edge_f=None, is_weighted=True, no_python=True)
>      45     print (embedding._method_name+':\n\tTraining time: %f' % (time() - t1))
>      46     # Evaluate on graph reconstruction
>
> /Users/jsgalan/Library/Python/2.7/lib/python/site-packages/gem-1.0.0-py2.7.egg/gem/embedding/node2vec.pyc in learn_embedding(self, graph, edge_f, is_weighted, no_python)
>      81         except Exception as e:
>      82             print(str(e))
> ---> 83             raise Exception('./node2vec not found. Please compile snap, place node2vec in the path and grant executable permission')
>      84         self._X = graph_util.loadEmbedding('tempGraph.emb')
>      85         t2 = time()
>
> Exception: ./node2vec not found. Please compile snap, place node2vec in the path and grant executable permission_

- when trying SDNE
_

> ---> 14 from gem.embedding.sdne     import SDNE
>      15
>      16 # File that contains the edges. Format: source target
>
> /Users/jsgalan/Library/Python/2.7/lib/python/site-packages/gem-1.0.0-py2.7.egg/gem/embedding/sdne.py in <module>()
>      26 from keras import backend as KBack
>      27
> ---> 28 from theano.printing import debugprint as dbprint, pprint
>      29 from time import time
>      30
>
> /Users/jsgalan/Library/Enthought/Canopy_64bit/User/lib/python2.7/site-packages/theano/__init__.py in <module>()
>      86
>      87
> ---> 88 from theano.configdefaults import config
>      89 from theano.configparser import change_flags
>      90
>
> /Users/jsgalan/Library/Enthought/Canopy_64bit/User/lib/python2.7/site-packages/theano/configdefaults.py in <module>()
>      45                      convert=floatX_convert,),
>      46              # TODO: see gh-4466 for how to remove it.
> ---> 47              in_c_key=True
>      48              )
>      49
>
> /Users/jsgalan/Library/Enthought/Canopy_64bit/User/lib/python2.7/site-packages/theano/configparser.pyc in AddConfigVar(name, doc, configparam, root, in_c_key)
>     279         if hasattr(root, name):
>     280             raise AttributeError('This name is already taken',
> --> 281                                  configparam.fullname)
>     282         configparam.doc = doc
>     283         configparam.in_c_key = in_c_key
>
> AttributeError: ('This name is already taken', 'floatX')

_

@palash1992 can you give me pointers to circumvent this errors?

Best
