### 2018-08-05 74acb3e
Usage simplifications
