+++ gem/embedding/sdne.py
@@ -25,7 +25,6 @@ import keras.regularizers as Reg
 from keras.optimizers import SGD, Adam
 from keras import backend as KBack
 
-from theano.printing import debugprint as dbprint, pprint
 from time import time
 
 
