### 2018-08-06 2b94804
Changed test file to Karate
