### 2018-08-07 59900ab
Made test.py consistent with readme
