### 2018-08-07 #876 @palash1992
@rougier : I will add that by the end of the day. I can add more examples as well.
