+++ gem/evaluation/visualize_embedding.py
@@ -22,7 +22,7 @@ def plot_embedding2D(node_pos, node_colors=None, di_graph=None):
         pos = {}
         for i in range(node_num):
             pos[i] = node_pos[i, :]
-        if node_colors:
+        if node_colors is not None:
             nx.draw_networkx_nodes(di_graph, pos,
                                    node_color=node_colors,
                                    width=0.1, node_size=100,
