+++ README.md
@@ -127,23 +127,22 @@ The output of the above execution is:
     Training time: 13.969912 MAP: 0.635860976597 preccision curve: [0.0, 0.0, 0.0, 0.25, 0.4]
 ```
 #### Visualization of Karate graph using Graph Factorization
-<img align="left" width="100" height="100" src="images/karate_graph_factor_sgd.png">
-<!-- ![karate_graph_factor_sgd](images/karate_graph_factor_sgd.png "Visualization of Karate graph using Graph Factorization") -->
+<img align="left" width="300" height="300" src="images/karate_graph_factor_sgd.png">
 
 #### Visualization of Karate graph using HOPE
-![karate_hope_gsvd](images/karate_hope_gsvd.png "Visualization of Karate graph using HOPE")
+<img align="left" width="300" height="300" src="images/karate_hope_gsvd.png">
 
 #### Visualization of Karate graph using Laplacian Eigenmaps
-![karate_lap_eigmap_svd](images/karate_lap_eigmap_svd.png "Visualization of Karate graph using Laplacian Eigenmaps")
+<img align="left" width="300" height="300" src="images/karate_lap_eigmap_svd.png">
 
 #### Visualization of Karate graph using Locally Linear Embedding
-![karate_lle_svd](images/karate_lle_svd.png "Visualization of Karate graph using Locally Linear Embedding")
+<img align="left" width="300" height="300" src="images/karate_lle_svd.png">
 
 #### Visualization of Karate graph using node2vec
-![karate_node2vec_rw](images/karate_node2vec_rw.png "Visualization of Karate graph using node2vec")
+<img align="left" width="300" height="300" src="images/karate_node2vec_rw.png">
 
 #### Visualization of Karate graph using SDNE
-![karate_sdne](images/karate_sdne.png "Visualization of Karate graph using SDNE")
+<img align="left" width="300" height="300" src="images/karate_sdne.png">
 
 ### Example 2
 Run the graph embedding methods on Stochastic Block Model graph and evaluate them on graph reconstruction and visualization. Please copy the gem/data/sbm.gpickle and gem/data/sbm_node_labels.pickle to the working directory:
@@ -234,23 +233,22 @@ The output of the above execution is:
     Training time: 667.998180 MAP: 0.9912109375 preccision curve: [1.0, 1.0, 1.0, 1.0, 1.0]
 ```
 #### Visualization of SBM using Graph Factorization
-<img align="left" width="100" height="100" src="images/sbm_graph_factor_sgd.png">
-<!-- ![sbm_graph_factor_sgd](images/sbm_graph_factor_sgd.png "Visualization of SBM using Graph Factorization") -->
+<img align="left" width="300" height="300" src="images/sbm_graph_factor_sgd.png">
 
 #### Visualization of SBM using HOPE
-![sbm_hope_gsvd](images/sbm_hope_gsvd.png "Visualization of SBM using HOPE")
+<img align="left" width="300" height="300" src="images/sbm_hope_gsvd.png">
 
 #### Visualization of SBM using Laplacian Eigenmaps
-![sbm_lap_eigmap_svd](images/sbm_lap_eigmap_svd.png "Visualization of SBM using Laplacian Eigenmaps")
+<img align="left" width="300" height="300" src="images/sbm_lap_eigmap_svd.png">
 
 #### Visualization of SBM using Locally Linear Embedding
-![sbm_lle_svd](images/sbm_lle_svd.png "Visualization of SBM using Locally Linear Embedding")
+<img align="left" width="300" height="300" src="images/sbm_lle_svd.png">
 
 #### Visualization of SBM using node2vec
[…]
