+++ README.md
@@ -127,22 +127,34 @@ The output of the above execution is:
     Training time: 13.969912 MAP: 0.635860976597 preccision curve: [0.0, 0.0, 0.0, 0.25, 0.4]
 ```
 #### Visualization of Karate graph using Graph Factorization
-<img align="left" width="300" height="300" src="images/karate_graph_factor_sgd.png">
+<p align="center">
+  <img width="300" height="300" src="images/karate_graph_factor_sgd.png">
+</p>
 
 #### Visualization of Karate graph using HOPE
-<img align="left" width="300" height="300" src="images/karate_hope_gsvd.png">
+<p align="center">
+<img width="300" height="300" src="images/karate_hope_gsvd.png">
+</p>
 
 #### Visualization of Karate graph using Laplacian Eigenmaps
-<img align="left" width="300" height="300" src="images/karate_lap_eigmap_svd.png">
+<p align="center">
+<img width="300" height="300" src="images/karate_lap_eigmap_svd.png">
+</p>
 
 #### Visualization of Karate graph using Locally Linear Embedding
-<img align="left" width="300" height="300" src="images/karate_lle_svd.png">
+<p align="center">
+<img width="300" height="300" src="images/karate_lle_svd.png">
+</p>
 
 #### Visualization of Karate graph using node2vec
-<img align="left" width="300" height="300" src="images/karate_node2vec_rw.png">
+<p align="center">
+<img width="300" height="300" src="images/karate_node2vec_rw.png">
+</p>
 
 #### Visualization of Karate graph using SDNE
-<img align="left" width="300" height="300" src="images/karate_sdne.png">
+<p align="center">
+<img width="300" height="300" src="images/karate_sdne.png">
+</p>
 
 ### Example 2
 Run the graph embedding methods on Stochastic Block Model graph and evaluate them on graph reconstruction and visualization. Please copy the gem/data/sbm.gpickle and gem/data/sbm_node_labels.pickle to the working directory:
@@ -233,22 +245,34 @@ The output of the above execution is:
     Training time: 667.998180 MAP: 0.9912109375 preccision curve: [1.0, 1.0, 1.0, 1.0, 1.0]
 ```
 #### Visualization of SBM using Graph Factorization
-<img align="left" width="300" height="300" src="images/sbm_graph_factor_sgd.png">
+<p align="center">
+<img width="300" height="300" src="images/sbm_graph_factor_sgd.png">
+</p>
 
 #### Visualization of SBM using HOPE
-<img align="left" width="300" height="300" src="images/sbm_hope_gsvd.png">
+<p align="center">
+<img width="300" height="300" src="images/sbm_hope_gsvd.png">
+</p>
 
 #### Visualization of SBM using Laplacian Eigenmaps
-<img align="left" width="300" height="300" src="images/sbm_lap_eigmap_svd.png">
+<p align="center">
+<img width="300" height="300" src="images/sbm_lap_eigmap_svd.png">
+</p>
 
 #### Visualization of SBM using Locally Linear Embedding
-<img align="left" width="300" height="300" src="images/sbm_lle_svd.png">
+<p align="center">
+<img width="300" height="300" src="images/sbm_lle_svd.png">
+</p>
 
 #### Visualization of SBM using node2vec
-<img align="left" width="300" height="300" src="images/sbm_node2vec_rw.png">
+<p align="center">
+<img width="300" height="300" src="images/sbm_node2vec_rw.png">
+</p>
 
 #### Visualization of SBM using SDNE
-<img align="left" width="300" height="300" src="images/sbm_sdne.png">
+<p align="center">
+<img width="300" height="300" src="images/sbm_sdne.png">
+</p>
 
 ## Cite
     @article{goyal2017graph,
