### 2018-08-09 9784db4
parameter explanation in readme
