+++ README.md
@@ -84,14 +84,20 @@ G = G.to_directed()
 
 models = []
 # You can comment out the methods you don't want to run
+# GF takes embedding dimension (d), maximum iterations (max_iter), learning rate (eta), regularization coefficient (regu) as inputs
 models.append(GraphFactorization(d=2, max_iter=100000, eta=1*10**-4, regu=1.0))
+# HOPE takes embedding dimension (d) and decay factor (beta) as inputs
 models.append(HOPE(d=4, beta=0.01))
+# LE takes embedding dimension (d) as input
 models.append(LaplacianEigenmaps(d=2))
+# LLE takes embedding dimension (d) as input
 models.append(LocallyLinearEmbedding(d=2))
+# node2vec takes embedding dimension (d),  maximum iterations (max_iter), random walk length (walk_len), number of random walks (num_walks), context size (con_size), return weight (ret_p), inout weight (inout_p) as inputs
 models.append(node2vec(d=2, max_iter=1, walk_len=80, num_walks=10, con_size=10, ret_p=1, inout_p=1))
-models.append(SDNE(d=2, beta=5, alpha=1e-5, nu1=1e-6, nu2=1e-6, K=3,n_units=[50, 15,], rho=0.3, n_iter=50, xeta=0.01,n_batch=500,
-                modelfile=['./intermediate/enc_model.json', './intermediate/dec_model.json'],
-                weightfile=['./intermediate/enc_weights.hdf5', './intermediate/dec_weights.hdf5']))
+# SDNE takes embedding dimension (d), seen edge reconstruction weight (beta), first order proximity weight (alpha), lasso regularization coefficient (nu1), ridge regreesion coefficient (nu2), number of hidden layers (K), size of each layer (n_units), number of iterations (n_ite), learning rate (xeta), size of batch (n_batch), location of modelfile and weightfile save (modelfile and weightfile) as inputs
+models.append(SDNE(d=2, beta=5, alpha=1e-5, nu1=1e-6, nu2=1e-6, K=3, n_units=[50, 15,], n_iter=50, xeta=0.01, n_batch=500,
+                modelfile=['enc_model.json', 'dec_model.json'],
+                weightfile=['enc_weights.hdf5', 'dec_weights.hdf5']))
 
 for embedding in models:
     print ('Num nodes: %d, num edges: %d' % (G.number_of_nodes(), G.number_of_edges()))
@@ -101,6 +107,9 @@ for embedding in models:
     print (embedding._method_name+':\n\tTraining time: %f' % (time() - t1))
     # Evaluate on graph reconstruction
     MAP, prec_curv, err, err_baseline = gr.evaluateStaticGraphReconstruction(G, embedding, Y, None)
+    #---------------------------------------------------------------------------------
+    print(("\tMAP: {} \t precision curve: {}\n\n\n\n"+'-'*100).format(MAP,prec_curv[:5]))
+    #---------------------------------------------------------------------------------
     # Visualize
     viz.plot_embedding2D(embedding.get_embedding(), di_graph=G, node_colors=None)
     plt.show()
@@ -109,23 +118,25 @@ for embedding in models:
 The output of the above execution is:
 ```
 ---------graph_factor_sgd:
-    Training time: 0.459782 MAP: 0.418189831501 preccision curve: [1.0, 0.5, 0.6666666666666666, 0.5, 0.4]
[…]
