### 2018-08-09 #876 @rougier
I still got some problems with missing files (filed an issue on the repo). I think you might want to have an example directory where you can put the `test_karate.py` and `test_sbm.py` files with the relevant data. Is there any reason why you put some data under `gem/data` ?  Overall, I think you might need to think of a better structure for your library. Everything's here but it is now mostly a matter of organization. If you look at the criterion for the review (at the start of this thread) it might help you in deciding what's the best structure.
