+++ tests/test_karate.py
@@ -5,10 +5,6 @@ gem/data/karate.edgelist to the working directory
 '''
 import matplotlib.pyplot as plt
 from time import time
-import networkx as nx
-try: import cPickle as pickle
-except: import pickle
-import numpy as np
 
 from gem.utils      import graph_util, plot_util
 from gem.evaluation import visualize_embedding as viz
@@ -24,31 +20,26 @@ from gem.embedding.sdne     import SDNE
 
 # File that contains the edges. Format: source target
 # Optionally, you can add weights as third column: source target weight
-# Copy the gem/data/sbm.gpickle and gem/data/sbm_node_labels.pickle 
-# to the working directory or specify the path to the file
-file_prefix = 'sbm.gpickle'
+# Copy the gem/data/karate.edgelist to the working directory or specify the path to the file
+edge_f = 'data/karate.edgelist'
 # Specify whether the edges are directed
 isDirected = True
 
 # Load graph
-G = nx.read_gpickle(file_prefix)
-node_colors = pickle.load(
-    open('sbm_node_labels.pickle', 'rb')
-)
-node_colors_arr = [None] * node_colors.shape[0]
-for idx in range(node_colors.shape[0]):
-    node_colors_arr[idx] = np.where(node_colors[idx, :].toarray() == 1)[1][0]
+G = graph_util.loadGraphFromEdgeListTxt(edge_f, directed=isDirected)
+G = G.to_directed()
 
 models = []
 # Load the models you want to run
-models.append(GraphFactorization(d=128, max_iter=1000, eta=1 * 10**-4, regu=1.0, data_set='sbm'))
-models.append(HOPE(d=256, beta=0.01))
-models.append(LaplacianEigenmaps(d=128))
-models.append(LocallyLinearEmbedding(d=128))
-models.append(node2vec(d=182, max_iter=1, walk_len=80, num_walks=10, con_size=10, ret_p=1, inout_p=1, data_set='sbm'))
-models.append(SDNE(d=128, beta=5, alpha=1e-5, nu1=1e-6, nu2=1e-6, K=3,n_units=[500, 300,], rho=0.3, n_iter=30, xeta=0.001,n_batch=500,
+models.append(GraphFactorization(d=2, max_iter=50000, eta=1 * 10**-4, regu=1.0))
+models.append(HOPE(d=4, beta=0.01))
+models.append(LaplacianEigenmaps(d=2))
+models.append(LocallyLinearEmbedding(d=2))
+models.append(node2vec(d=2, max_iter=1, walk_len=80, num_walks=10, con_size=10, ret_p=1, inout_p=1))
+models.append(SDNE(d=2, beta=5, alpha=1e-5, nu1=1e-6, nu2=1e-6, K=3,n_units=[50, 15,], rho=0.3, n_iter=50, xeta=0.01,n_batch=100,
                 modelfile=['enc_model.json', 'dec_model.json'],
                 weightfile=['enc_weights.hdf5', 'dec_weights.hdf5']))
+
 # For each model, learn the embedding and evaluate on graph reconstruction and visualization
 for embedding in models:
     print ('Num nodes: %d, num edges: %d' % (G.number_of_nodes(), G.number_of_edges()))
@@ -59,9 +50,9 @@ for embedding in models:
     # Evaluate on graph reconstruction
     MAP, prec_curv, err, err_baseline = gr.evaluateStaticGraphReconstruction(G, embedding, Y, None)
     #---------------------------------------------------------------------------------
-    print(("\tMAP: {} \t preccision curve: {}\n\n\n\n"+'-'*100).format(MAP,prec_curv[:5]))
+    print(("\tMAP: {} \t preccision curve: {}\n\n\n\n"+'-'*100).format(MAP,prec_curv))
[…]
