+++ tests/test_karate.py
@@ -20,7 +20,6 @@ from gem.embedding.sdne     import SDNE
 
 # File that contains the edges. Format: source target
 # Optionally, you can add weights as third column: source target weight
-# Copy the gem/data/karate.edgelist to the working directory or specify the path to the file
 edge_f = 'data/karate.edgelist'
 # Specify whether the edges are directed
 isDirected = True
