### 2018-08-09 #876 @palash1992
@rougier I have now created a "tests" folder with the required files following the instructions. This should resolve the issues.
