### 2018-08-09 1610816
fixed pickle incompatibility between python 2 and 3
