### 2018-08-09 7e69bd7
Networkx version compatibility
