### 2018-08-09 3c4dacc
printing top 5 prec values
