### 2018-08-22 #876 @rougier
Ok. You'll need to add SNAP in the list of dependencies (and it might be good to add links for each of the dependency)
