### 2018-08-22 #876 @palash1992
@rougier: My idea for not adding node2vec to dependencies was based on the premise that some users may not need to run node2vec and thus should not be restricted by having library dependency. Adding the fallback is a good alternative. At this point I can make one of the below three modifications:

1. Add an argument in test.py specifying the method to be tested. Thus, the program would be run as python test.py -m all (or) python test.py -m hope,lap,gf. If the user is interested in node2vec, as specified in the installation instructions he/she can specify it in the argument.
2. Add SNAP as the dependency.
3. Add https://github.com/aditya-grover/node2vec as the fallback and dependency of gensim in the library. If the user installs SNAP, the library will automatically take the installation.

Please let me know which of these are you in favor of.
