### 2018-08-23 #876 @rougier
This is where I get lost and I think the culprit is the lack of documentation ;) Is node2vec necessary only to run the tests or is it necessary to run the software? If this is only for the test, then it's fine for me. I can run the karate test and  probably @jsgalan can run the other one. However, it would be good in the documentation to explain (for example) the karate example (what is the expected output, what does it mean, etc) and to explain that second example requires the SNAP library. In the end, it would be only a matter of changing the README.

Also, maybe the `tests` directory could be renamed `examples` since content is actually illustrating usage.

While checking the paper, I think you're missing a DOI for Belkin et al. 2002.
