### 2018-08-24 f65112d
Renamed tests to examples and optionally running node2vec
