### 2018-08-24 dc2b13c
Added description to run test and test results
