+++ paper.bib
@@ -10,12 +10,13 @@
   publisher={American Association for the Advancement of Science}
 }
 
-@inproceedings{belkin2002laplacian,
-  title={Laplacian eigenmaps and spectral techniques for embedding and clustering},
+@inproceedings{belkin2003laplacian,
+  title={Laplacian eigenmaps for dimensionality reduction and data representation},
   author={Belkin, Mikhail and Niyogi, Partha},
-  booktitle={Advances in Neural Information Processing System},
-  pages={585--591},
-  year={2002}
+  booktitle={Neural Computation},
+  doi={10.1162/089976603321780317},
+  pages={1373--1396},
+  year={2003}
 }
 
 @inproceedings{ahmed2013distributed,
