### 2018-08-24 2a5d219
Update paper.md
