### 2018-08-25 #876 @jsgalan
@rougier after the pip install networkx==1.11 some parts of the tests where smooth. you might want to try.
