### 2018-08-28 #876 @palash1992
@rougier : Let me know if the issue still persists
