### 2018-08-30 #876 @rougier
@palash1992 Just did and it works (with some warnings though)
