### 2018-08-30 #876 @palash1992
@rougier : Perfect! I also made a release. Let me know if there is anything else needed from my side.
