### 2018-08-30 #876 @danielskatz
@rougier - I see there are still quite a few checkboxes left empty.  Please go through them when you have a chance.
