### 2018-08-31 #876 @rougier
Just did. The **Functionality documentation** and **Community guidelines** could be improved a bit but it does not prevent publication. Thanks @palash1992 for taking all my comments into account.
