### 2018-08-31 #876 @danielskatz
great, thanks @rougier !

@palash1992  - please make an archive of the current state of the repo in Zenodo or similar, and let me know the DOI of the archive, so that can move forward in publishing this submission.
