### 2020-01-15 #876 @Laya1363
Hi every one.
I want to apply GEMs in my graph. in order to succeed I started to download all needed packages and for stating point I wanted to test the example programs in GEM repository but continuously I encounter different issues while running. can any one help? for example while running sbm example I recieve this issue:

Using TensorFlow backend.
Traceback (most recent call last):
  File "C:/Laya/MasterThesis/programming/test1.py", line 46, in <module>
    G = nx.read_gpickle(file_prefix)
  File "<c:\laya\softwares\python36_x64\lib\site-packages\decorator.py:decorator-gen-174>", line 2, in read_gpickle
  File "c:\laya\softwares\python36_x64\lib\site-packages\networkx\utils\decorators.py", line 220, in _open_file
    result = func(*new_args, **kwargs)
  File "c:\laya\softwares\python36_x64\lib\site-packages\networkx\readwrite\gpickle.py", line 101, in read_gpickle
    return pickle.load(path)
_pickle.UnpicklingError: invalid load key, '\x0a'.

I am really confiused. can anyone help ?
