### 2020-01-15 #876 @danielskatz
👋 @Laya1363 - you are asking in the wrong place - this issue is for the JOSS review of the software, which has now been published, not for use of the software.  Please open an issue in the software repository - https://github.com/palash1992/GEM
